Local Global Switch - Parent Constraints and Weighting is now implemented
Standard weight switch, using the same model as the FK/IK Switch code...
However, the renaming was upgraded a little, by capturing the actual
name of the node generated (parC) rather than the name we built and
assigned to it. This defends against accidental duplicate names.

This could be upgraded on the FKIKSwitch if needed